+++ src/requisition-batch-approval/requisition-batch-approval.controller.js
@@ -233,7 +233,10 @@
 
                 // Using slice to make copy of array, so scope changes at end only
                 requisitionBatchApproveFactory.batchApprove(vm.requisitions.slice())
-                .then(handleApprove, handleApprove);
+                .then(handleApprove, handleApprove)
+                .catch(function() {
+                    loadingModalService.close();
+                });
             });
         }
 
@@ -316,7 +319,6 @@
 
                 // Reload state to display page without approved notifications and to update outdated ones
                 $state.go($state.current, {errors: errors, ids: requisitionIds.join(',')}, {reload: true});
-                loadingModalService.close();
 
             } else {