+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ===========
 
 * [OLMIS-3201](https://openlmis.atlassian.net/browse/OLMIS-3201): Removed stockout days validation and added popover for stockout days field.
+* [OLMIS-3294](https://openlmis.atlassian.net/browse/OLMIS-3294): Added loading modal after the approval is finished.
 
 5.1.1 / 2017-09-01
 ===========